Add specific Exceptions classes

+++ src/accessors/BaseAccessor.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -7,6 +8,32 @@ using System.Web.Script.Serialization;
 
 namespace GitLabSharp
 {
+   /// <summary>
+   /// Exception class for non-specific issues.
+   /// </summary>
+   public class GitLabSharpException : Exception
+   {
+      public GitLabSharpException(string url, string error)
+         : base(String.Format("Error occurred with URL \"{0}\": {1}", url, error))
+      {
+      }
+   }
+
+   /// <summary>
+   /// Exception class for System.Net.WebException class of issues.
+   /// </summary>
+   public class GitLabRequestException : Exception
+   {
+      public GitLabRequestException(string url, string method, System.Net.HttpStatusCode code)
+         : base(String.Format("GitLab returned error code {0} on requesting URL \"{1}\" with method {2}",
+            code.ToString(), url, method))
+      {
+         Code = code;
+      }
+
+      public System.Net.HttpStatusCode Code { get; }
+   }
+
    /// <summary>
    /// Base class for all Accessors - classes, that run Http Requests to access different kinds of data
    /// </summary>
@@ -23,7 +50,7 @@ namespace GitLabSharp
       /// </summary>
       internal T Get<T>(string url)
       {
-         return Serializer.Deserialize<T>(Client.Get(url));
+         return Serializer.Deserialize<T>(safeRequest(url, "GET"));
       }
 
       /// <summary>
@@ -31,7 +58,7 @@ namespace GitLabSharp
       /// </summary>
       internal T Post<T>(string url)
       {
-         return Serializer.Deserialize<T>(Client.Post(url));
+         return Serializer.Deserialize<T>(safeRequest(url, "POST"));
       }
 
       /// <summary>
@@ -39,7 +66,7 @@ namespace GitLabSharp
       /// </summary>
       internal T Put<T>(string url)
       {
-         return Serializer.Deserialize<T>(Client.Put(url));
+         return Serializer.Deserialize<T>(safeRequest(url, "PUT"));
       }
 
       /// <summary>
@@ -47,7 +74,44 @@ namespace GitLabSharp
       /// </summary>
       internal void Delete(string url)
       {
-         Client.Delete(url);
+         safeRequest(url, "DELETE");
+      }
+
+      /// <summary>
+      /// Executes a request but converts System.Net.WebException into GitLabSharpException
+      /// </summary>
+      private string safeRequest(string url, string method)
+      {
+         string response = String.Empty;
+         try
+         {
+            if (method == "GET")
+            {
+               response = Client.Get(url);
+            }
+            else if (method == "POST")
+            {
+               response = Client.Post(url);
+            }
+            else if (method == "PUT")
+            {
+               response = Client.Put(url);
+            }
+            else if (method == "DELETE")
+            {
+               response = Client.Delete(url);
+            }
+            else
+            {
+               Debug.Assert(false);
+            }
+         }
+         catch (System.Net.WebException ex)
+         {
+            var exResponse = ((System.Net.HttpWebResponse)ex.Response);
+            throw new GitLabRequestException(url, method, exResponse.StatusCode);
+         }
+         return response;
       }
 
       protected JavaScriptSerializer Serializer = new JavaScriptSerializer();
@@ -55,3 +119,4 @@ namespace GitLabSharp
       internal string BaseUrl { get; }
    }
 }
+
